+++ EarthAsylumConsulting/Helpers/ipUtil.class.php
@@ -154,7 +154,7 @@ class ipUtil
 	 * Compares two IPv6 addresses.
 	 * In case a subnet is given, it checks if it contains the request IP.
 	 *
-	 * @author David Soria Parra <dsp at php dot net>
+	 * author David Soria Parra <dsp at php dot net>
 	 *
 	 * @see https://github.com/dsp/v6tools
 	 *